# crates/manta-cli/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Manta API",
    "description": "REST API for managing CSM/OpenCHAMI HPC clusters via manta.",
    "contact": {
      "name": "Manuel Sopena Ballesteros",
      "email": "msopena@cscs.ch"
    },
    "license": {
      "name": ""
    },
    "version": "2.0.0-beta.64"
  },
  "servers": [
    {
      "url": "/api/v2",
      "description": "manta API v2"
    }
  ],
  "paths": {
...
    "/hardware-clusters/{target}/configuration": {
      "post": {
        "tags": [
          "hardware"
        ],
        "summary": "`POST /api/v2/hardware-clusters/{target}/configuration` — pin or unpin nodes between clusters by hardware pattern.",
        "operationId": "apply_hw_configuration",
        "parameters": [
          {
            "name": "target",
            "in": "path",
            "description": "Target cluster name",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Manta-Site",
            "in": "header",
            "description": "Name of the target cluster (matches a site configured in the server).",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApplyHwConfigurationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Configuration applied or preview",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/hardware-clusters/{target}/members": {
      "post": {
        "tags": [
          "hardware"
        ],
        "summary": "`POST /api/v2/hardware-clusters/{target}/members` — move nodes matching a hardware pattern into a cluster.",
        "operationId": "add_hw_component",
        "parameters": [
          {
            "name": "target",
            "in": "path",
            "description": "Target cluster name",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Manta-Site",
            "in": "header",
            "description": "Name of the target cluster (matches a site configured in the server).",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddHwComponentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Members added or preview",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "hardware"
        ],
        "summary": "`DELETE /api/v2/hardware-clusters/{target}/members` — move nodes back to parent cluster by hardware pattern.",
        "operationId": "delete_hw_component",
        "parameters": [
          {
            "name": "target",
            "in": "path",
            "description": "Target cluster name",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Manta-Site",
            "in": "header",
            "description": "Name of the target cluster (matches a site configured in the server).",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteHwComponentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Members removed or preview",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/images": {
      "get": {
        "tags": [
          "images"
        ],
        "summary": "GET /images — list IMS images, oldest-first (newest last).",
        "operationId": "get_images",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "description": "Exact IMS image ID; returns just that image when set.",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "pattern",
            "in": "query",
            "description": "Glob pattern matched against image name; applied server-side\n(`service::image::get_images`). Invalid glob returns 400.",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "since",
            "in": "query",
            "description": "ISO-8601 lower bound — only images created at or after this\npoint. Applied server-side; IMS cannot filter by date itself.",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "until",
            "in": "query",
            "description": "ISO-8601 upper bound — only images created at or before this\npoint. Applied server-side; IMS cannot filter by date itself.",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Cap on the number of images returned (the newest N; listed oldest\nfirst, newest last).",
            "required": false,
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "format": "int32",
              "minimum": 0
            }
          },
          {
            "name": "X-Manta-Site",
            "in": "header",
            "description": "Name of the target cluster (matches a site configured in the server).",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of images",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {}
                }
              }
            }
          },
          "400": {
            "description": "Malformed since/until, or since after until",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "images"
        ],
        "summary": "`DELETE /api/v2/images` — delete IMS images by ID; validates only when `dry_run=true`.",
        "operationId": "delete_images",
        "parameters": [
          {
            "name": "ids",
            "in": "query",
            "description": "Comma-separated list of IMS image IDs to delete.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "dry_run",
            "in": "query",
            "description": "When true, validate deletion eligibility without removing\nanything.",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "X-Manta-Site",
            "in": "header",
            "description": "Name of the target cluster (matches a site configured in the server).",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Images deleted or validation result",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "AddHwComponentRequest": {
        "type": "object",
        "description": "Request body for `POST /api/v2/hardware-clusters/{target}/members`.\n\nMoves nodes matching `pattern` out of `parent_cluster` and into\nthe path-level target cluster. The reverse direction lives in\n[`DeleteHwComponentRequest`]; both operations are also expressible\nin unified form via [`ApplyHwConfigurationRequest`] with\n[`HwClusterMode::Pin`] vs [`HwClusterMode::Unpin`].",
        "required": [
          "parent_cluster",
          "pattern"
        ],
        "properties": {
          "create_hsm_group": {
            "type": "boolean",
            "description": "Create the target HSM group if it does not already exist."
          },
          "dry_run": {
            "type": "boolean",
            "description": "When true, return the planned changes without modifying group\nmembership."
          },
          "parent_cluster": {
            "type": "string",
            "description": "Source HSM group that donates nodes matching `pattern`."
          },
          "pattern": {
            "type": "string",
            "description": "Hardware component pattern used to select which nodes to move."
          }
        }
      },
...
      "ApplyHwConfigurationRequest": {
        "type": "object",
        "description": "Request body for\n`POST /api/v2/hardware-clusters/{target}/configuration`.",
        "required": [
          "parent_cluster",
          "pattern"
        ],
        "properties": {
          "create_target_hsm_group": {
            "type": "boolean",
            "description": "Create the target HSM group if absent (default `true`)."
          },
          "delete_empty_parent_hsm_group": {
            "type": "boolean",
            "description": "Delete the parent HSM group if it becomes empty (default `true`)."
          },
          "dry_run": {
            "type": "boolean",
            "description": "When true, return the planned changes without modifying group\nmembership."
          },
          "mode": {
            "$ref": "#/components/schemas/HwClusterMode",
            "description": "Whether to pin nodes into the target cluster or unpin them back\nto the parent. Defaults to `Pin`."
          },
          "parent_cluster": {
            "type": "string",
            "description": "Source (parent) HSM group supplying nodes."
          },
          "pattern": {
            "type": "string",
            "description": "Hardware component pattern selecting which nodes to pin/unpin."
          }
        }
      },
...
      "DeleteHwComponentRequest": {
        "type": "object",
        "description": "Request body for `DELETE /api/v2/hardware-clusters/{target}/members`.\n\nReverse of [`AddHwComponentRequest`]: moves nodes matching\n`pattern` out of the path-level target cluster and back to\n`parent_cluster`.",
        "required": [
          "parent_cluster",
          "pattern"
        ],
        "properties": {
          "delete_hsm_group": {
            "type": "boolean",
            "description": "Delete the target HSM group if it becomes empty after the\noperation."
          },
          "dry_run": {
            "type": "boolean",
            "description": "When true, return the planned changes without modifying group\nmembership."
          },
          "parent_cluster": {
            "type": "string",
            "description": "Destination HSM group that receives nodes moved out of the\ntarget cluster."
          },
          "pattern": {
            "type": "string",
            "description": "Hardware component pattern used to select which nodes to move\nback."
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "description": "Standard JSON error body returned by all failed endpoints.",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string",
            "description": "Human-readable explanation of the failure. Never includes\nstack traces, credentials, or internal type names."
          }
        }
      },
...
      "HwClusterMode": {
        "type": "string",
        "description": "Whether the hw cluster operation moves nodes into the target (Pin) or\nreleases them back (Unpin).",
        "enum": [
          "pin",
          "unpin"
        ]
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}